% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cermine-package.R
\docType{package}
\name{cermine-package}
\alias{cermine}
\alias{cermine-package}
\title{cermine: Scrape raw or agreggated data for Maturita and JPZ exams by CZVV}
\description{
CZVV neither publishes a REST API nor provides exam data in easy-to-process JSON or XML formats. Instead, the data are published in separate XLSX files that are difficult to access programmatically because the links are dynamically generated by an ASP.NET server. This package provides a simple R API that allows you to gather links to all available exam data files and returns them in a tidy table.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/netique/cermine}
  \item \url{https://netique.github.io/cermine/}
  \item Report bugs at \url{https://github.com/netique/cermine/issues}
}

}
\author{
\strong{Maintainer}: Jan Netik \email{netikja@gmail.com} (\href{https://orcid.org/0000-0002-3888-3203}{ORCID})

}
\keyword{internal}
