% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mine.R
\name{mine_links}
\alias{mine_links}
\title{Get the links to files from the CZVV website}
\usage{
mine_links(
  projects = NULL,
  years = NULL,
  data_type = "item",
  force = FALSE,
  base_url = default_base_url()
)
}
\arguments{
\item{projects}{Projects to extract the links for. If \code{NULL} (the default),
all projects available are used. Possible values are \verb{Maturitní zkoušky}
and \verb{Jednotná přijímací zkouška} (but that may be a subject to change).
You'll get the up-to-date options listed if you type in incorrect
option(s). To get the current options, you may also call
\code{get_exam_events()} first and use the \code{project} column from the result.}

\item{years}{Years to extract the links for. If \code{NULL} (the default), all
years available are used. You'll get the up-to-date options listed if you
type in incorrect option(s). To get the current options, you may also call
\code{get_exam_events()} first and use the \code{year} column from the result. Note
that if you opt for some impossible combination of project and year, the
function will inform you.}

\item{data_type}{Type of the data to return the links for. \code{item} (the
default) extracts only item data for every student, \code{aggregated} extracts
only aggregated data, and \code{both} returns both types.}

\item{force}{Force a refresh of the data, even if cached. Defaults to
\code{FALSE}.}

\item{base_url}{Base URL of the CERMAT website. Defaults to
\url{https://vysledky.cermat.cz/statistika/}.}
}
\description{
From the default URL, (\url{https://vysledky.cermat.cz/statistika/}), the
function first gets all the project-year combinations you can possibly set
using the form fields \code{Projekt} and \code{Rok}, then proceeds to fetch the HTMLs
returned by the server. Finally, it extracts the links to the files and
returns a \code{tibble} with project, year, data type (either "item" or
"aggregated") and the link to the data file you can use to download the data.
To download the files, and store them is a handy file structure, use
\code{{cermine}}'s function \code{download_data()} on the result of \code{mine_links()}.
}
\details{
Note that \code{Rok} is a slight misnomer -- it actually refers to the exam
"event" (usually it comes in year/season format, e.g. "2025 - Jaro" or
"2025 - Podzim"), however, for "Jednotné přijímací zkoušky", all terms are
presented under single year option. Note further that naming of \code{Rok} field
options is not consistent.
}
