% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cermine-package.R
\docType{package}
\name{cermine-package}
\alias{cermine}
\alias{cermine-package}
\title{cermine: Scrape item-level or agreggated data for Maturita and JPZ exams by CZVV}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/netique/cermine}
  \item \url{https://netique.github.io/cermine/}
  \item Report bugs at \url{https://github.com/netique/cermine/issues}
}

}
\author{
\strong{Maintainer}: Jan Netik \email{netikja@gmail.com} (\href{https://orcid.org/0000-0002-3888-3203}{ORCID})

}
\keyword{internal}
