% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{mine_links}
\alias{mine_links}
\title{Get the links to files from the CZVV website}
\usage{
mine_links(
  projects = NULL,
  years = NULL,
  data_type = "item",
  base_url = default_base_url()
)
}
\arguments{
\item{data_type}{Type of the data to return the links for. \code{item} (the default)
extracts only item data for every student, \code{aggregated} extracts only aggregated data, and \code{both}
returns both types.}

\item{base_url}{Base URL of the CERMAT website. Defaults to
\url{https://vysledky.cermat.cz/statistika/}.}
}
\description{
From the default URL,
(\url{https://vysledky.cermat.cz/statistika/}), the function first gets all the
project-year combinations you can possibly set using the form fields
\code{Projekt} and \code{Rok}, then proceeds to fetch the HTMLs returned by the
server. Finally, it extracts the links to the files and returns a table. To download the files,
you can use \code{{cermine}}'s function \code{download_data()} on the result of \code{mine_links()}.
}
\details{
Note that \code{Rok} is a slight misnomer -- it actually refers to the exam
"event" (usually it comes in year/season format, e.g. "2025 - Jaro" or "2025 - Podzim"),
however, for "Jednotné přijímací zkoušky", all terms are
presented under single year option. Note further that naming of \code{Rok} field options
is not consistent.
}
